Destroy flammable items when in fire or lava
The value of the `flammable` group of the item will determine how
frequently will the check for igniters (fire/lava) be performed.

It will be done every `20/flammable` seconds, non-flammable items
won't check for igniters.

Once an igniter is found, the value of its `igniter` + `lava` groups
will be used to determine how long will it take for the item to
disappear, `20/(igniter+lava)` seconds.